+++ src/core/models/general-configuration.ts
@@ -1,0 +1,7 @@
+export interface GeneralConfiguration {
+    configurationFilePath: string;
+}
+
+export const defaultGeneralConfiguration: GeneralConfiguration = {
+    configurationFilePath: './import-sorter.json'
+};