+++ src/core/models/index.ts
@@ -1,6 +1,7 @@
 export * from './import-element';
 export * from './import-node';
 export * from './import-string-configuration';
+export * from './general-configuration';
 export * from './sort-configuration';
 export * from './import-element-group';
 export * from './import-element-sort-result';