+++ CHANGELOG.md
@@ -1,9 +1,8 @@
 # [0.0.2]
 ### New features
 * added support for trailing commas. `importSorter.importStringConfiguration.trailingComma` is an enum which can be `always`, `multiLine`, or `none`.
-`always` - will always append trailing comma.
-`multiLine` - will append comma if import line is broken to multiple lines.
-`none` - will not append comma(`This one is a default setting`)
+
+* added configuration file support. Configuration can be read from json file. The file is specified by `importSorter.generalConfiguration.configurationFilePath` setting.
 
 # Initial version 0.0.1
 First prototype is released